Uploading the files of LUNG_CANCER_PREDICTION
The primary aim of this project is to develop and evaluate machine learning models that can predict lung cancer based on a range of clinical and behavioural features. By using various machine learning techniques, such as Logistic Regression, Random Forest, and Support Vector Machines (SVM), the project seeks to identify which algorithm performs best in predicting lung cancer. The project’s goal is to build
 a robust, accurate model capable of identifying individuals who are at a higher risk of developing lung cancer, thereby enabling early intervention and improving the chances of successful treatment. Additionally, the project aims to explore the influence of different factors, such as smoking and chronic diseases, in predicting lung cancer and to assess the model's performance through metrics such as accuracy, precision, recall, and ROC-AUC scores.